add standard file structure, remove mocks, and demo platform
