+++ Services/include/communication/communication_service.h
@@ -26,6 +26,52 @@
 #define MIN_FREQ = 2200
 #define MAX_FREQ = 2290
 
+#define TC_COMMUNICATION_SERVICE 10
+typedef enum {
+  S_GET_TEMP = 0,
+  S_GET_FREQ = 1,
+  S_GET_CONTROL = 2,
+  S_GET_ENCODER = 3,
+  S_GET_PA_POWER = 4,
+  S_GET_STATUS = 5,
+  S_GET_TR = 6,
+  S_GET_BUFFER = 7,
+  S_GET_HK = 8,
+  S_SOFT_RESET = 9,
+  S_GET_FULL_STATUS = 10,
+  S_SET_FREQ = 11,
+  S_SET_CONTROL = 12,
+  S_SET_ENCODER = 13,
+  S_SET_PA_POWER = 14,
+  S_GET_CONFIG = 15,
+  S_SET_CONFIG = 16
+} Sband_Subtype;
+
+typedef enum {
+  UHF_SET_STAT_CONTROL = 20,
+  UHF_SET_FREQ,
+  UHF_SET_PIPE_TIMEOUT,
+  UHF_SET_BEACON_T,
+  UHF_SET_AUDIO_T,
+  UHF_SET_PARAMS,  // Merge last 4
+  UHF_RESTORE_DEFAULT,
+  UHF_LOW_PWR,
+  UHF_SET_DEST,
+  UHF_SET_SRC,
+  UHF_SET_MORSE,
+  UHF_SET_MIDI,
+  UHF_SET_BEACON_MSG,
+  UHF_SET_I2C,
+  UHF_WRITE_FRAM,
+  UHF_SECURE,
+  UHF_GET_FULL_STAT,  // 0-244,249,250
+  UHF_GET_CALL_SIGN,  // 245-246
+  UHF_GET_MORSE,
+  UHF_GET_MIDI,
+  UHF_GET_BEACON_MSG,
+  UHF_GET_FRAM,
+} UHF_Subtype;
+
 SAT_returnState communication_service_app(csp_packet_t* pkt);
 
 SAT_returnState start_communication_service(void);
