+++ Services/source/util/service_utilities.c
@@ -16,12 +16,10 @@
  * @author upSat, Andrew Rooney
  * @date 2020-06-06
  */
-#include "service_utilities.h"
+#include "util/service_utilities.h"
 
 #include <stdarg.h>
 #include <stdio.h>
-
-#include "time_management_service.h"
 
 void ex2_log(const char *format, ...) {
   //  va_list arg;
